Breaking down Layer.swift into a directory  (#102)
Fix https://github.com/tensorflow/swift-apis/issues/69.
